- use rollup to generate umd module - add babel to transpile es6 code - update test to use built index.js in dist/
