+++ package.json
@@ -2,11 +2,12 @@
   "name": "arabic-reshaper",
   "version": "1.1.0",
   "description": "Reshapes Arabic words to work in non-supported environments.",
-  "main": "index.js",
+  "main": "dist/index.js",
   "directories": {
     "test": "test"
   },
   "scripts": {
+    "build": "rollup -c rollup.config.js",
     "test": "mocha",
     "coverage": "istanbul cover _mocha --"
   },
@@ -21,11 +22,15 @@
   },
   "homepage": "https://github.com/louy/Javascript-Arabic-Reshaper#readme",
   "devDependencies": {
+    "babel-core": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
     "chai": "^3.4.0",
     "codacy-coverage": "^1.1.3",
     "codecov.io": "^0.1.6",
     "istanbul": "^0.4.0",
     "mocha": "^2.3.3",
-    "mocha-junit-reporter": "^1.8.0"
+    "mocha-junit-reporter": "^1.8.0",
+    "rollup": "^0.50.0",
+    "rollup-plugin-babel": "^3.0.2"
   }
 }
