+++ rollup.config.js
@@ -1,0 +1,15 @@
+import babel from 'rollup-plugin-babel';
+
+export default {
+  input: 'src/index.js',
+  output: {
+    file: 'dist/index.js',
+    format: 'umd',
+    name: 'ArabicReshaper',
+  },
+  plugins: [
+    babel({
+      exclude: 'node_modules/**'
+    })
+  ]
+}
